+++ asm/hwc_to_chw.go
@@ -3,6 +3,6 @@
 
 package asm
 
-func Hwc2Chw(output []float32, input []float32, height int, width int) {
+func Hwc2Chw(output []uint8, input []uint8, height int, width int) {
 	nativeHwc2Chw(output, input, height, width)
 }
