+++ asm/hwc_to_chw_amd64.go
@@ -15,7 +15,7 @@ func __hwc2chw(result unsafe.Pointer, input unsafe.Pointer, height uint64, width
 //xxx go:noescape
 // func ___hwc2chw2(width uint32, height uint32, mean unsafe.Pointer, input unsafe.Pointer, result unsafe.Pointer)
 
-func Hwc2Chw(output []float32, input []float32, height int, width int) {
+func Hwc2Chw(output []uint8, input []uint8, height int, width int) {
 	// println("height  = ", uint32(height))
 	// println("width  = ", uint32(width))
